Change the representation of `RelocatableValue`
It used to hold a SegmentedAddr.  This caused issues in semantics that require
the ability to perform addition and subtraction operations on RelocatableValues.
Addition or subtraction operations on SegmentedAddrs are not well defined,
because these operations can push a value from one segment into another, and
there is no way for them to compute the new segment.

The new representation is just a MemWord, whose interpretation is now decided
later (in `transferValue'` in AbsState).