Add a little state to keep track if we are translating an AVX instruciton.
Writes to an XMM register may or may not affect to corresponding YMM
register, depending on the type of instruction we are using.
With SSE instruction, the upper bits of YMM are left as is, while with
AVX the upper bits are set to 0.

The `avxMode` field of the monad keeps track of what insturction we
are translating.