symbolic: Rewrite the translation of macaw calls into crucible
The former strategy was to represent macaw calls using a macaw-specific
MacawCall statement, which was interpreted by a call handler (which took
registers+memory as input and produced new registers+memory as an output).  This
worked for cases where the callee had a summary, but did not allow for
simulating the called function inline.  Moreover, the OverrideSim monad doesn't
admit recursive calls in this context (we can make the call, but we can't get
the final simulator state out, which we would need to implement a call handler
in macaw-symbolic).

The new strategy is to translate macaw calls into two separate statements:

1. A `LookupFunctionHandle` call, which returns a Crucible FunctionHandle, and
2. A normal Crucible `Call`

The interpretation of LookupFunctionHandle has the full register+memory state
available, and can inspect the IP to determine which function has been
called (and provide the necessary FunctionHandle, which will be interpreted by
Crucible in the standard way).  Note that the handler is in IO, so client code
can translate functions being simulated into Crucible on-demand.