Revise the GlobalMap API
The GlobalMap is mapping from virtual addresses computed by a program to the
corresponding logical address in the LLVM memory model during symbolic
simulation.  It is needed because addresses in binaries are computed from
bitvectors, which are not valid pointers in the LLVM memory model.

This change turns the GlobalMap from a Data.Map into a function, which is more
flexible and allows for a wider range of possible implementations of this
functionality, especially implementations that introduce numerous disjoint
segments for the original binary contents.