Improve error reporting in TH generated semantics
There was an error case in function interpretation in the TH generated
code (when a function couldn't be evaluated for a given operand).  This
shouldn't happen for well-formed code, but can be a problem when macaw finds
invalid code that happens to decode as a real instruction (with an invalid
operand).

The old code called error, which caused macaw to fail with a hard stop.  This
commit changes the call to be to `fail` instead, which fires off an exception
that is properly handled (giving us a ClassifyFailure instead).