Turn fail-stop errors into macaw translation failures
Previously, macaw-ppc (and macaw-arm) would call `error` if there were no
semantics available for a decoded instruction.  This was useful during initial
development, but it is a problem for deployment.  Now just turn missing
semantics into TranslationErrors, which appear as block terminators in macaw IR.

This will require more diligence in monitoring TranslationErrors for patterns
that need to be addressed.