[refinement] Track all solutions and ensure latest function is analyzed.
For each discovered transfer refinement, the Discovery process must be
re-started from scratch for the particular target function in
question.  The implication is that subsequent refinements for the same
nominal function must use the newer version, and that all previous
refinements must be re-applied each time the function Discovery
process is repeated.

These changes ensure that both of these occur during the refinement
iterations.