[x86_symbolic] Use stable value for generating x86 register name.
Relying on the underlying Show instance representation for the
register name is more brittle, especially if the Show becomes a
structural representation (e.g. "X86_GP 14") that introduces spaces
and other elements that can cause problems.  Generate the register
name in a more reliable manner for the symbolic variable use case
here.