Generalize function argument analysis.
This patch focuses on function argument analysis, but includes some
other cleanups.

The main changes are to add additional comments and cleanups to the
function argument analysis code.  This also extends the analysis so
that we can annotate the types of some of the functions and use those
types during analysis.

As part of this we tighten the PLTStub checking, and clean up the
elfloader in some minor ways.