No-op change to re-trigger Travis.
The previous commits were valid, but github requires Travis builds to
complete successfully before allowing a merge to master.  The Travis
builds did complete, but somehow the communication between Travis and
Github failed, so Github still is waiting for the Travis build.
Pushing this minor change to re-trigger Travis builds so that this
branch can be merged to master.