Attach externally-learned block addresses to DiscoveryFunInfo
After we learn new block targets from the SMT solver, we now have a way to
attach them to the discovered function as a side mapping.  The mapping is first
used to populate the discovery frontier in a fresh run of the code discovery.
It is then saved so that downstream code can properly interpret the
ClassifyFailures that have been resolved.

This feature could also be used to seed macaw with information from other code
discovery engines.