Fix a solver state scoping error
We need to run the symbolic execution in a fresh solver frame so that it doesn't
pollute the environment for when we ask for IP models.

This commit has a bit of tracing added that will be removed eventually.

The current problem in PowerPC examples is that r2 isn't populated, which means
that the solver just cooks up completely unconstrained models of the IP, which
aren't right.  This can be fixed by augmenting the initial solver state.

The current x86 examples all work fine in normal macaw now.