Populate known initial state when symbolically simulating a path
This information comes from the abstract domains computed by macaw during code
discovery.  In some cases, it can identify constants in registers at the start
of a block.  We populate our initial states with that information to further
constrain the search.  This is really important for handling the TOC in PowerPC.

This has the solver returning a real value discovered for some PPC examples.
Unfortunately, the results are still under-constrained.