Remove an unnecessary allocation
This allocation was never really used, but it added an extra "live" allocation
in the memory model.  Since the memory model couldn't tell where a symbolic
pointer might point, it added mux cases to cover this extra region.
Unfortunately, the extra region had entirely symbolic contents, which meant that
the resulting IPs calculated from it were totally unconstrained.  This allowed
the solver to generate many invalid IP models.

It looks like the region was added to provide a "default" pointer for the
address mapping code; this wasn't actually necessary since that code only
affects the mapping of pointers with symbolic region numbers (which don't occur
in this use case).  We can provide the NULL pointer as the default with no harm.