Add extra information needed to really plumb refinement results into macaw
We need to include reasons that each new block was discovered (easy).  We also
need to provide values for the abstract domains and jump bounds for the
discovered blocks.  This is harder - right now we just provide some conservative
defaults.

This change modifies how we identify unknown transfers.  We do not flag jumps
we've already resolved as unknown; to do so, we need to consult the side map we
introduce into the DiscoveryFunInfo.

Experiments show that the conservative approximations of the AbsBlockState and
JumpBounds are not good enough.  After refining the function and re-running
discovery, macaw fails to identify the return (instead marking it as a
ClassifyFailure).  macaw-refinement then attempts to refine the return, which is
under-constrained.  We will have to compute proper abstractions.