Fix injection of externally-discovered control flow into macaw base
This new approach steps in after a block has been flagged as ending in a
classification failure.  It just takes the externally-discovered results and
injects them as intra-procedural jump targets for the block (leaving the
terminator as a ClassifyFailure).  We compute reasonable abstract processor
states and jump bounds for the new blocks.  After this, the existing intra-jump
analysis code propagates enough information to make everything just work out.

The previous approach was problematic because it didn't have enough information
available to make suitable abstract states, which ended up breaking the analysis
of calls and returns (because the abstract state for the link register/stack
went to Top, because we set all of our initial states for the injected blocks to
Top).  Note that this commit has to back out the changes in mkFunState, because
they clobber the function discovery state with some invalid abstract states.

As an alternative to this design choice, we could probably make the
incorporation of externally-provided branch resolution into a Classifier in the
existing classifier framework.  I didn't go that route because it would always
have to be last (or first, depending on policy if we want it to override macaw's
normal findings).  That would be an unfortunate informal policy to have to
enforce.  The approach taken in this patch makes that obvious by just
interposing at the last feasible time.