Use the SimpleBackend for symbolic execution
Start up a new online connection (a low-level one, not the OnlineBackend from
Crucible) to generate IP models.  There was a theory that the extra bookkeeping
by the OnlineBackend might be causing trouble, but that seems to not be the
case.

With a bit more diagnosis, it seems like the problem is that reads from symbolic
pointers make a mux across all possible allocations.  One way to minimize the
problem is to actually use a single allocation for all of memory instead so that
we never need to make that mux.  The single allocation will be sparse (i.e.,
there will be regions with no data).  We'll need to make sure that we can
generate safety checks to ensure that no pointers read from/written to touch
those invalid regions.

There will also be questions about how to handle stacks and the heap, as a fully
symbolic pointer could touch any of those, too (bringing us back to the muxed
memory region problem).