Significantly revise the memory model for macaw-symbolic
Instead of using one memory allocation for each loadable segment, use a single
allocation for all of memory.

Multiple allocations largely prevents us from taking advantage of the benefits
of the symbolic memory model: that the SMT solver does all of the work.  If we
use multiple allocations, we have to mux across all of the allocations if we
have a symbolic block ID.

There is still some work to do with this change:
- The mux tree isn't created for multiple allocations if there are extras (there
would be with dynamic allocation, unless we make a symbolic memory allocator...)
- We need to add safety assertions to make sure that symbolic reads don't touch
unmapped memory
- If we expose the base pointer, we can compute where the stack should live in
this memory space

There are still some open questions about how this should all be done, and if
there need to be multiple possible configurations.