Revise the machine code memory model
This commit adds the necessary bounds checks on pointers to the static memory
region.  It handles the cases where the assertions must be conditional (due to
e.g., symbolic pointers and conditional writes).

This actually simplifies the pointer mapping code, which can mostly go away.
There is a new function argument that controls what pointer validity predicate
is asserted at each use site.  The suggested memory
model (Macaw.Symbolic.Memory) provides an implementation, but it could be
overridden if desired (e.g., if the use case in question doesn't bother to check
the assertions).

This commit also fixes what I believe was a bug in tryGlobPtr.  Previously, that
function only translated pointers that had a constant 0 BlockID.  That isn't
right: it actually needs to (conditionally) translate pointers with symbolic
block ids (e.g., creating an LLVMPointerMux).  This commit fixes that and only
skips translating pointers that have a concrete non-zero BlockID.