Revise the macaw-symbolic memory model again
This was an experiment with updating the initial array contents directly instead
of representing concrete values with assumptions.  Unfortunately, it was not
successful right now because doing so kills all of our SMT solvers.  This commit
leaves the experiments commented out so that we can try as the solvers are
fixed; in theory, it should be more efficient the other way.