Fix the memory protection assertion in macaw-symbolic
Since we have a single allocation for static allocations taken from a binary, we
have regions of the underlying array that are "unmapped".  We represent that by
generating a safety condition that any access has to be within the bounds of the
mapped regions.

The condition was incorrect before: it omitted the first byte of each
region (and extended one past the end).  This commit swaps it to be correct.