Rebuild the test harness
It is now a lot simpler.  It doesn't have all of the features of the original.
It now only checks the output of refinement against expected output files.  They
are not golden-style tests - instead, they just state how many blocks the
interesting function should have and what their addresses are.

This new test harness shares most of its code with the run-refinement binary.