Fix the macaw-x86-symbolic tests
Adapts the test suite to use the new memory model initialization
infrastructure.  The commit changes the target of the test to a
statically-linked binary instead of an object file.  Object file support does
not currently work, as the provided memory model only supports a single memory
region.  Macaw addresses put object files in a separate memory region, which is
not yet accounted for in the memory model code.