Re-implement timeouts, record failures programatically
Re-implements timeouts safely so that we don't end up killing threads from the
scheduler thread pool (which ends up just breaking).  We now run our own async
timeout thread that calls killSolver.

This change also records refinement failures in a structured way so that we can
inspect them and report them to the user, rather than just spilling them to
stderr.  A consequence of this improved tracking is that we can keep track of
past failures and avoid failing again on the same intractable problems.  That
tracking lets this change also iterate the refinement process until nothing
changes.